refacotr: move signal handling and pprof to grace package
