+++ weed/util/grace/pprof.go
@@ -1,4 +1,4 @@
-package util
+package grace
 
 import (
 	"os"
